add: dwm config to xinitrc
the first task uses the grep command with the -q flag to search for "exec dwm" in the .xinitrc file. If the string is found, the command returns a rc value of 0, which is considered a success. If the string is not found or an error occurs, the rc value is non-zero.

The register keyword is used to store the output of the shell command in the xinitrc_file variable. The stdout attribute contains the output of the command (which is empty if the string is not found), and the rc attribute contains the return code.